effect フィールドエフェクトデータコンバートディレクトリ とりあえずPLから影と煙絵を持ってくる fldmapdata/makefile effect追加 arc_tool.lst ARCID_FLDEFF追加 make_prog_files fldeff_shadow.c追加 field_effect.c,h 参照関数追加、タスク追加時に初期化関数を読んでいないバグ修正 field_player_grid.c とりあえずかげエフェクトを自機に付けてみる fieldmap.c,h フィールドエフェクト登録を追加 fieldmap_ctrl_grid.c デバッグ出力無効 fldeff_shadow.c,h 影エフェクト用ソースファイル fldmmdl.c,h 引数zone_idの型をu16に

+++ prog/src/field/fldmmdl.c
@@ -3940,7 +3940,7 @@ static void FldMMdl_ChangeOBJAlies(
  * @retval	int		TRUE=ìØàÍÅBFALSE=ìØàÍÇ≈ÇÕÇ»Ç¢
  */
 //--------------------------------------------------------------
-BOOL FLDMMDL_CheckSameID( const FLDMMDL * fmmdl, u16 obj_id, int zone_id )
+BOOL FLDMMDL_CheckSameID( const FLDMMDL * fmmdl, u16 obj_id, u16 zone_id )
 {
 	if( FLDMMDL_CheckStatusBit(fmmdl,FLDMMDL_STABIT_USE) == 0 ){
 		return( FALSE );
@@ -3975,7 +3975,7 @@ BOOL FLDMMDL_CheckSameID( const FLDMMDL * fmmdl, u16 obj_id, int zone_id )
  */
 //--------------------------------------------------------------
 BOOL FLDMMDL_CheckSameIDCode(
-		const FLDMMDL * fmmdl, u16 code, u16 obj_id, int zone_id )
+		const FLDMMDL * fmmdl, u16 code, u16 obj_id, u16 zone_id )
 {
 	if( FLDMMDL_CheckStatusBit(fmmdl,FLDMMDL_STABIT_USE) ){
 		return( FALSE );
